+++ longevity/src/main/scala/longevity/exceptions/subdomain/IndexDuplicatesKeyException.scala
@@ -4,10 +4,10 @@ import emblem.TypeKey
 
 /** thrown on attempt to construct a
  * [[longevity.subdomain.Subdomain Subdomain]] with a
- * [[longevity.subdomain.PType PType]] that has two or more keys or indexes
+ * [[longevity.subdomain.PType PType]] that has a key and an index
  * defined over the same properties
  */
-class DuplicateKeyOrIndexException(
+class IndexDuplicatesKeyException(
   val pTypeKey: TypeKey[_])
 extends SubdomainException(
   s"PType ${pTypeKey.name} has duplicate keys or indexes")