+++ longevity/src/test/scala/longevity/integration/noTranslation/WithNoTranslation.scala
@@ -9,6 +9,5 @@ case class WithNoTranslation(
 object WithNoTranslation extends PType[WithNoTranslation] {
   object props {
   }
-  object keys {
-  }
+  val keySet = emptyKeySet
 }