+++ longevity/src/test/scala/longevity/integration/noTranslation/WithNoTranslationList.scala
@@ -9,6 +9,5 @@ case class WithNoTranslationList(
 object WithNoTranslationList extends PType[WithNoTranslationList] {
   object props {
   }
-  object keys {
-  }
+  val keySet = emptyKeySet
 }