+++ longevity/src/test/scala/longevity/integration/noTranslation/WithNoTranslationOption.scala
@@ -9,6 +9,5 @@ case class WithNoTranslationOption(
 object WithNoTranslationOption extends PType[WithNoTranslationOption] {
   object props {
   }
-  object keys {
-  }
+  val keySet = emptyKeySet
 }