+++ longevity/src/test/scala/longevity/integration/subdomain/foreignKey/WithForeignKey.scala
@@ -1,20 +1,10 @@
 package longevity.integration.subdomain.foreignKey
 
-import longevity.subdomain.PType
+import longevity.subdomain.annotations.persistent
 
+@persistent(
+  keySet = Set(key(props.id)),
+  indexSet = Set(index(props.associated)))
 case class WithForeignKey(
   id: WithForeignKeyId,
   associated: AssociatedId)
-
-object WithForeignKey extends PType[WithForeignKey] {
-  object props {
-    val id = prop[WithForeignKeyId]("id")
-    val associated = prop[AssociatedId]("associated")
-  }
-  object keys {
-    val id = key(props.id)
-  }
-  object indexes {
-    val associated = index(props.associated)
-  }
-}