+++ longevity/src/test/scala/longevity/unit/subdomain/PTypeSpec.scala
@@ -2,7 +2,6 @@ package longevity.unit.subdomain
 
 import longevity.exceptions.subdomain.ptype.MultiplePartitionKeysForPType
 import longevity.exceptions.subdomain.ptype.NoPropsForPTypeException
-import longevity.exceptions.subdomain.ptype.NoKeysForPTypeException
 import longevity.exceptions.subdomain.ptype.PartitionKeyForDerivedPTypeException
 import longevity.subdomain.PType
 import org.scalatest.FlatSpec
@@ -26,9 +25,7 @@ object PTypeSpec {
     object props {
       val username = prop[Username]("username")
     }
-    object keys {
-      val username = partitionKey(props.username)
-    }
+    val keySet = Set(partitionKey(props.username))
   }
 
   case class Email(email: String) extends KeyVal[EmailedUser]
@@ -39,9 +36,7 @@ object PTypeSpec {
     object props {
       val email = prop[Email]("email")
     }
-    object keys {
-      val email = partitionKey(props.email)
-    }
+    val keySet = Set(partitionKey(props.email))
   }
 
 }
@@ -55,8 +50,7 @@ class PTypeSpec extends FlatSpec with GivenWhenThen with Matchers {
 
   it should "throw exception when the `propSet` is not overridden, and there is no `object props`" in {
     object User extends PType[User] {
-      object keys {
-      }
+      val keySet = emptyKeySet
     }
     intercept[NoPropsForPTypeException[_]] {
       User.propSet
@@ -67,8 +61,7 @@ class PTypeSpec extends FlatSpec with GivenWhenThen with Matchers {
     object User extends PType[User] {
       object props {
       }
-      object keys {
-      }
+      val keySet = emptyKeySet
     }
     User.propSet should equal (Set())
   }
@@ -79,66 +72,9 @@ class PTypeSpec extends FlatSpec with GivenWhenThen with Matchers {
         val username = prop[Username]("username")
         val email = prop[Email]("email")
       }
-      object keys {
-      }
+      val keySet = emptyKeySet
     }
     User.propSet should equal (Set(User.props.username, User.props.email))
-  }
-
-  it should "throw exception when the `keySet` is not overridden, and there is no `object keys`" in {
-    object User extends PType[User] {
-      object props {
-      }
-    }
-    intercept[NoKeysForPTypeException[_]] {
-      User.keySet
-    }
-  }
-
-  it should "produce an empty `keySet` when `object keys` is empty" in {
-    object User extends PType[User] {
-      object props {
-      }
-      object keys {
-      }
-    }
-    User.keySet should equal (Set())
-  }
-
-  it should "produce an empty `keySet` when `object keys` is non-empty but holds no keys" in {
-    object User extends PType[User] {
-      object props {
-      }
-      object keys {
-        val x = 7
-      }
-    }
-    User.keySet should equal (Set())
-  }
-
-  it should "produce an empty `keySet` when `object keys` is holds no keys of the right type" in {
-    object User extends PType[User] {
-      object props {
-      }
-      object keys {
-        val x = Blog.keys.uri
-      }
-    }
-    User.keySet should equal (Set())
-  }
-
-  it should "produce a non-empty `keySet` when `object keys` holds keys of the right type" in {
-    object User extends PType[User] {
-      object props {
-        val username = prop[Username]("username")
-        val email = prop[Email]("email")
-      }
-      object keys {
-        val username = key(props.username)
-        val email = key(props.email)
-      }
-    }
-    User.keySet should equal (Set(User.keys.username, User.keys.email))
   }
 
   it should "throw exception if more than one partition key is defined" in {
@@ -147,10 +83,7 @@ class PTypeSpec extends FlatSpec with GivenWhenThen with Matchers {
         val username = prop[Username]("username")
         val email = prop[Email]("email")
       }
-      object keys {
-        val username = partitionKey(props.username)
-        val email = partitionKey(props.email)
-      }
+      val keySet = Set(partitionKey(props.username), partitionKey(props.email))
     }
     intercept[MultiplePartitionKeysForPType[_]] {
       User.partitionKey
@@ -168,64 +101,9 @@ class PTypeSpec extends FlatSpec with GivenWhenThen with Matchers {
     object User extends PType[User] {
       object props {
       }
-      object keys {
-      }
+      val keySet = emptyKeySet
     }
     User.indexSet should equal (Set())
-  }
-
-  it should "produce an empty `indexSet` when `object indexes` is empty" in {
-    object User extends PType[User] {
-      object props {
-      }
-      object keys {
-      }
-      object indexes {
-      }
-    }
-    User.indexSet should equal (Set())
-  }
-
-  it should "produce an empty `indexSet` when `object indexes` is non-empty but holds no indexes" in {
-    object User extends PType[User] {
-      object props {
-      }
-      object keys {
-      }
-      object indexes {
-        val x = 7
-      }
-    }
-    User.indexSet should equal (Set())
-  }
-
-  it should "produce an empty `indexSet` when `object indexes` is holds no indexes of the right type" in {
-    object User extends PType[User] {
-      object props {
-      }
-      object keys {
-      }
-      object indexes {
-        val x = Blog.index(Blog.props.uri)
-      }
-    }
-    User.indexSet should equal (Set())
-  }
-
-  it should "produce a non-empty `indexSet` when `object indexes` holds indexes of the right type" in {
-    object User extends PType[User] {
-      object props {
-        val username = prop[String]("username")
-        val email = prop[Email]("email")
-      }
-      object keys {
-      }
-      object indexes {
-        val username = index(props.username)
-        val email = index(props.email)
-      }
-    }
-    User.indexSet should equal (Set(User.indexes.username, User.indexes.email))
   }
 
   behavior of "PType.toString"
@@ -234,8 +112,7 @@ class PTypeSpec extends FlatSpec with GivenWhenThen with Matchers {
     object User extends PType[User] {
       object props {
       }
-      object keys {
-      }
+      val keySet = emptyKeySet
     }
     User.toString should equal ("PType[User]")
   }