+++ longevity/src/test/scala/longevity/unit/subdomain/annotations/MPropsSpec.scala
@@ -30,7 +30,9 @@ object MPropsSpec {
     // @mprops can't work on a type that is declared in the same object
 
     // help ensure macro application doesnt mess up companion object
-    @mprops class WithCompanion extends PType[Foo]
+    @mprops class WithCompanion extends PType[Foo] {
+      override val keySet = emptyKeySet
+    }
 
     object WithCompanion { val y = 7 }
   }